feat: Add transformOnAutoComplete option
I added the option "transformOnAutoComplete" because I am using this
module for path auto-completion.
The transformOnAutoComplete option allows me to append a '/' to the
auto-completed user input in case the path is a directory, but it
could certainly be used for other things as well.
In case the transformOnAutoComplete option is used, another auto-completion
attempt is made using the new, transformed input.